Fix prototype (&) to auto-unwrap REF to CODE
When a subroutine has prototype (&) and receives \(&code), Perl
automatically unwraps the double reference to get the code reference.
This is standard Perl behavior.

Added unwrapping logic in handleCodeReferenceArgument() to detect
OperatorNode(\) containing ListNode with OperatorNode(&) and unwrap
it to just the & node.

Fixes tests 58-62 in comp/proto.t where a_sub \(&tmp_sub_1) was
failing with 'Not a CODE reference' error.