Fix 'when' keyword parsing to respect 'use feature switch'
The 'when' keyword should only be treated as a reserved keyword when
'use feature "switch"' is active. Without this feature enabled, 'when'
should be allowed as a bareword, particularly as a hash key.

Changes:
- StatementResolver.java: Add feature flag checks for 'given', 'when',
  and 'default' keywords (similar to 'try' and 'class')
- ListParser.java: Add 'when' to the list of keywords that should be
  treated as barewords when followed by '=>' (fat comma)

This fixes parsing errors in ExifTool's XMP.pm where 'when' is used
as a hash key: when => { %dateTimeInfo, Groups => { 2 => 'Time' } }

Tested with ExifTool 13.44 - XMP-related commands now parse correctly.